requirements.txt
appium-python-client==3.2.1
selenium==4.22.0
python-dotenv==1.0.1
requests==2.32.3
rapidfuzz==3.9.7
pillow==11.3.0